AddonsSecurityBeans and @AutoConfigureAddonsSecurity
Use a single bean name for
- ServletSecurityBeans and ReactiveSecurityBeans: AddonsSecurityBeans
- @AutoConfigureAddonsSecurity{Webmvc|Weblux}{Jwt|Introspecting}: @AutoConfigureAddonsSecurity
