+++ archetypes/spring-addons-archetypes-webflux-multimodule/src/main/resources/archetype-resources/__api-artifactId__/src/test/java/ControllerTest.java
@@ -8,12 +8,12 @@ import java.lang.annotation.Target;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
 import org.springframework.context.annotation.Import;
 
-import com.c4_soft.springaddons.security.oauth2.test.webflux.jwt.AutoConfigureAddonsSecurityWebfluxJwt;
+import com.c4_soft.springaddons.security.oauth2.test.webflux.jwt.AutoConfigureAddonsSecurity;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @WebFluxTest
-@AutoConfigureAddonsSecurityWebfluxJwt
+@AutoConfigureAddonsSecurity
 @Import({ SampleApi.WebSecurityConfig.class, EnableSpringDataWebSupportTestConf.class })
 public @interface ControllerTest {
