+++ archetypes/spring-addons-archetypes-webmvc-singlemodule/src/main/resources/archetype-resources/src/test/java/ControllerTest.java
@@ -8,12 +8,12 @@ import java.lang.annotation.Target;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.Import;
 
-import com.c4_soft.springaddons.security.oauth2.test.webmvc.jwt.AutoConfigureAddonsSecurityWebmvcJwt;
+import com.c4_soft.springaddons.security.oauth2.test.webmvc.jwt.AutoConfigureAddonsSecurity;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @WebMvcTest
-@AutoConfigureAddonsSecurityWebmvcJwt
+@AutoConfigureAddonsSecurity
 @Import({ SampleApi.WebSecurityConfig.class, EnableSpringDataWebSupportTestConf.class })
 public @interface ControllerTest {
