+++ samples/README.md
@@ -8,7 +8,7 @@ Samples for different security scenari, with **configuration and unit tests** fo
 - granted authorities retrieved from the token or from an external source (JPA repo in the sample but could be a web-service)
 - usage of test annotations or "fluent API" (MockMvc request post-processors and WebTestClient mutators)
 
-All sample but `webmvc-keycloakauthenticationtoken` make usage of [spring-addons-webmvc-jwt-resource-server](https://github.com/ch4mpy/spring-addons/blob/master/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java) or [spring-addons-webflux-jwt-resource-server](https://github.com/ch4mpy/spring-addons/blob/master/webflux/spring-addons-webflux-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/reactive/ReactiveSecurityBeans.java) `@AutoConfiguration`.
+All sample but `webmvc-keycloakauthenticationtoken` make usage of [spring-addons-webmvc-jwt-resource-server](https://github.com/ch4mpy/spring-addons/blob/master/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java) or [spring-addons-webflux-jwt-resource-server](https://github.com/ch4mpy/spring-addons/blob/master/webflux/spring-addons-webflux-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/reactive/AddonsSecurityBeans.java) `@AutoConfiguration`.
 
 As a consequence there are 3 sources of configuration for each sample:
 - `application.properties` files
