+++ samples/tutorials/resource-server_with_introspection/README.md
@@ -113,11 +113,11 @@ import org.springframework.context.annotation.Import;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.OpenIdClaims;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.WithMockBearerTokenAuthentication;
 import com.c4_soft.springaddons.security.oauth2.test.mockmvc.MockMvcSupport;
-import com.c4_soft.springaddons.security.oauth2.test.mockmvc.introspecting.AutoConfigureAddonsSecurityWebmvcIntrospecting;
+import com.c4_soft.springaddons.security.oauth2.test.mockmvc.introspecting.AutoConfigureAddonsSecurity;
 import com.c4soft.springaddons.tutorials.ResourceServerWithOAuthenticationApplication.WebSecurityConfig;
 
 @WebMvcTest(controllers = GreetingController.class)
-@AutoConfigureAddonsSecurityWebmvcIntrospecting
+@AutoConfigureAddonsSecurity
 @Import(WebSecurityConfig.class)
 class GreetingControllerTest {
