+++ samples/webflux-oauthentication/src/test/java/com/c4_soft/springaddons/samples/webflux_oidcauthentication/GreetingControllerAnnotatedTest.java
@@ -23,12 +23,12 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 
-import com.c4_soft.springaddons.security.oauth2.config.reactive.ReactiveSecurityBeans;
+import com.c4_soft.springaddons.security.oauth2.config.reactive.AddonsSecurityBeans;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.OpenIdClaims;
 import com.c4_soft.springaddons.security.oauth2.test.annotations.WithMockJwtAuth;
 import com.c4_soft.springaddons.security.oauth2.test.webflux.AddonsWebfluxTestConf;
 import com.c4_soft.springaddons.security.oauth2.test.webflux.WebTestClientSupport;
-import com.c4_soft.springaddons.security.oauth2.test.webflux.jwt.AutoConfigureAddonsSecurityWebfluxJwt;
+import com.c4_soft.springaddons.security.oauth2.test.webflux.jwt.AutoConfigureAddonsSecurity;
 
 import reactor.core.publisher.Mono;
 
@@ -36,8 +36,8 @@ import reactor.core.publisher.Mono;
  * @author Jérôme Wacongne &lt;ch4mp&#64;c4-soft.com&gt;
  */
 @WebFluxTest(GreetingController.class)
-@AutoConfigureAddonsSecurityWebfluxJwt
-@Import({ SampleApi.WebSecurityConfig.class, ReactiveSecurityBeans.class, AddonsWebfluxTestConf.class })
+@AutoConfigureAddonsSecurity
+@Import({ SampleApi.WebSecurityConfig.class, AddonsSecurityBeans.class, AddonsWebfluxTestConf.class })
 class GreetingControllerAnnotatedTest {
 
 	@MockBean
