+++ webflux/spring-addons-webflux-introspecting-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/reactive/AddonsSecurityBeans.java
@@ -65,7 +65,7 @@ import reactor.core.publisher.Mono;
 @AutoConfiguration
 @Slf4j
 @Import(SpringAddonsSecurityProperties.class)
-public class ReactiveSecurityBeans {
+public class AddonsSecurityBeans {
 
 	/**
 	 * Hook to override security rules for all path that are not listed in "permit-all". Default is isAuthenticated().
