+++ webflux/spring-addons-webflux-jwt-resource-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.c4_soft.springaddons.security.oauth2.config.reactive.ReactiveSecurityBeans
+com.c4_soft.springaddons.security.oauth2.config.reactive.AddonsSecurityBeans
