+++ webflux/spring-addons-webflux-jwt-test/src/main/java/com/c4_soft/springaddons/security/oauth2/test/webflux/jwt/AutoConfigureAddonsSecurity.java
@@ -1,4 +1,4 @@
-package com.c4_soft.springaddons.security.oauth2.test.mockmvc.introspecting;
+package com.c4_soft.springaddons.security.oauth2.test.webflux.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,11 +8,11 @@ import java.lang.annotation.Target;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 
 import com.c4_soft.springaddons.security.oauth2.config.SpringAddonsSecurityProperties;
-import com.c4_soft.springaddons.security.oauth2.config.reactive.ReactiveSecurityBeans;
+import com.c4_soft.springaddons.security.oauth2.config.reactive.AddonsSecurityBeans;
 import com.c4_soft.springaddons.security.oauth2.test.webflux.AddonsWebfluxTestConf;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@ImportAutoConfiguration({ SpringAddonsSecurityProperties.class, ReactiveSecurityBeans.class, AddonsWebfluxTestConf.class })
-public @interface AutoConfigureAddonsSecurityWebfluxIntrospecting {
+@ImportAutoConfiguration({ SpringAddonsSecurityProperties.class, AddonsSecurityBeans.class, AddonsWebfluxTestConf.class })
+public @interface AutoConfigureAddonsSecurity {
 }
