+++ webmvc/spring-addons-webmvc-jwt-resource-server/README.MD
@@ -40,7 +40,7 @@ public class SampleApi {
 
 	@EnableGlobalMethodSecurity(prePostEnabled = true)
 	public static class WebSecurityConfig {
-	  // browse com.c4_soft.springaddons.security.oauth2.config.synchronised.ServletSecurityBeans
+	  // browse com.c4_soft.springaddons.security.oauth2.config.synchronised.AddonsSecurityBeans
 	  // for auto-configuration you can override here
 	}
 }
@@ -72,19 +72,19 @@ com.c4-soft.springaddons.security.statless-sessions=true
 ```
 
 ## How the magic happens
-This lib uses the new spring-boot (2.7 and above) auto-configuration: as per [`META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports`](https://github.com/ch4mpy/spring-addons/blob/master/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports) content, [`ServletSecurityBeans`](https://github.com/ch4mpy/spring-addons/blob/3fff5abc17a3594235d35ca15696cca30eb99f62/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L79) is auto-configured. This configuration class does two things:
+This lib uses the new spring-boot (2.7 and above) auto-configuration: as per [`META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports`](https://github.com/ch4mpy/spring-addons/blob/master/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports) content, [`AddonsSecurityBeans`](https://github.com/ch4mpy/spring-addons/blob/3fff5abc17a3594235d35ca15696cca30eb99f62/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L79) is auto-configured. This configuration class does two things:
 - `@Import({ SpringAddonsSecurityProperties.class })`: this enables `com.c4-soft.springaddons.security.*` properties
 - define all required web-security beans
 
 ## Configuration overrides
-All beans declared in [`ServletSecurityBeans`](https://github.com/ch4mpy/spring-addons/blob/3fff5abc17a3594235d35ca15696cca30eb99f62/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L79) are decorated with `@ConditionalOnMissingBean`. This means you can override any component just by defining bean of the same type in your app. Do not hesitate to open sources to get inspiration.
+All beans declared in [`AddonsSecurityBeans`](https://github.com/ch4mpy/spring-addons/blob/3fff5abc17a3594235d35ca15696cca30eb99f62/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L79) are decorated with `@ConditionalOnMissingBean`. This means you can override any component just by defining bean of the same type in your app. Do not hesitate to open sources to get inspiration.
 
 Here are the beans you might like to override:
-- [`ExpressionInterceptUrlRegistryPostProcessor`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L93): define security policy for non "permit-all" routes. Default requires users to be authenticated (not anonymous).
-- [`Jwt2ClaimSetConverter`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L210): Override to have something else than `OpenidClaimSet` in `OAuthentication<T>`. For instance a specialized implementation parsing OpenID private claims your authorization-server provides or even something just OAuth2 (not OpenID).
-- [`Jwt2AuthoritiesConverter`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L198): switch authorities converter to something completely at your hand (for instance to retrieve authorities from a service rather than token claims, or just because prefix and case transformation are not enough).
-- [`SynchronizedJwt2AuthenticationConverter<T>`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L183): switch from `OAuthentication<T>` to any other `Authentication` implementation of your choice.
-- [`HttpSecurityPostProcessor`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/ServletSecurityBeans.java#L105): post-process HttpSecurity after it was pre-configured according to `com.c4-soft.springaddons.security.*` properties. This enables to **change anything configured in this lib** in case the four preceding were not enough.
+- [`ExpressionInterceptUrlRegistryPostProcessor`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L93): define security policy for non "permit-all" routes. Default requires users to be authenticated (not anonymous).
+- [`Jwt2ClaimSetConverter`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L210): Override to have something else than `OpenidClaimSet` in `OAuthentication<T>`. For instance a specialized implementation parsing OpenID private claims your authorization-server provides or even something just OAuth2 (not OpenID).
+- [`Jwt2AuthoritiesConverter`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L198): switch authorities converter to something completely at your hand (for instance to retrieve authorities from a service rather than token claims, or just because prefix and case transformation are not enough).
+- [`SynchronizedJwt2AuthenticationConverter<T>`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L183): switch from `OAuthentication<T>` to any other `Authentication` implementation of your choice.
+- [`HttpSecurityPostProcessor`](https://github.com/ch4mpy/spring-addons/blob/0063225a1c181282fc6519db9fc277c187c0a5ed/webmvc/spring-addons-webmvc-jwt-resource-server/src/main/java/com/c4_soft/springaddons/security/oauth2/config/synchronised/AddonsSecurityBeans.java#L105): post-process HttpSecurity after it was pre-configured according to `com.c4-soft.springaddons.security.*` properties. This enables to **change anything configured in this lib** in case the four preceding were not enough.
 
 You should not override those 2 (you can, but if you have to, please open a ticket, there's something I might have missed):
 - `JwtIssuerAuthenticationManagerResolver`: enables mutli-tenancy.
