+++ webmvc/spring-addons-webmvc-jwt-resource-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.c4_soft.springaddons.security.oauth2.config.synchronised.ServletSecurityBeans
+com.c4_soft.springaddons.security.oauth2.config.synchronised.AddonsSecurityBeans
